Another round of Juno repo fixes
The previous fix for the EOL Juno RDO repository didn't last long as the RPM file for the repo is now gone.
This new fix copies the repo file itself everywhere (postscripts and containers). Note that it doesn't do a GPG check to avoid having to copy the key too, which would have been a bit cumbersome in the postscripts.

I grep'd through the sources to locate all references to the Juno repo. I don't think I have missed anything, but your never know.