Keep latest docker image tag for controller install
Commit 9c339dce862e18726cfb0ce67941e42ad27410a7 added the Trinity revision number as a tag for the freshly-built Docker images. When a tag is specified, docker build doesn't add the default latest tag anymore.

Controller postscripts pull the latest-tagged images, so since that commit those don't exist anymore (except if there is a pre-existing image already tagged as latest) and controller installs break.

This patch keeps the revision tag and restores the latest tag.

Associated ticket: #750

Note: The easiest way to do that turned out to be a second docker build with no tag specified. As the intermediate results aren't removed after the first build and the source files didn't change, the second build doesn't do anything and basically only tags the image from the first build.